Split out stubEndpoints()
