+++ cypress/tsconfig.json
@@ -5,6 +5,7 @@
   },
   "exclude": [],
   "include": [
-    "../__tests__/frontend/**/*"
+    "../__tests__/frontend/**/*",
+    "../__tests__/test-utils/stubEndpoints.ts"
   ]
 }
