+++ tsconfig.json
@@ -11,5 +11,8 @@
     "noUnusedParameters": true,
     "strict": true
   },
-  "exclude": ["__tests__/frontend/**/*"]
+  "exclude": [
+    "__tests__/frontend/**/*",
+    "__tests__/test-utils/stubEndpoints.ts"
+  ]
 }
